refactor: clean up files + file structure
