+++ src/components/.storybook/Styles/Styles.module.scss
@@ -1,4 +1,4 @@
-@import '../../styles/colors';
+@import '../../../styles/colors';
 
 // Export color variables for use in Webpack.
 :export {
