+++ src/components/AsyncTable/index.ts
@@ -1,0 +1,4 @@
+import AsyncTable from './AsyncTable';
+
+export type { AsyncTableProps, Column, SearchKey } from './AsyncTable';
+export default AsyncTable;
