+++ src/components/Button/index.ts
@@ -1,0 +1,4 @@
+import Button from './Button';
+
+export type { ButtonProps } from './Button';
+export default Button;
