+++ src/components/Card/index.ts
@@ -1,0 +1,4 @@
+import Card from './Card';
+
+export type { AsyncText, CardProps } from './Card';
+export default Card;
