+++ src/components/DropdownMenu/index.ts
@@ -1,0 +1,4 @@
+import DropdownMenu from './DropdownMenu';
+
+export type { DropdownItem, DropdownMenuProps } from './DropdownMenu';
+export default DropdownMenu;
