+++ src/components/Input/index.tsx
@@ -5,8 +5,6 @@ import { AriaTextFieldProps } from '@react-types/textfield';
 import TextInput from './Input';
 import NumberInput from './NumberInput';
 
-import './Input.scss';
-
 export interface InputProps extends Omit<AriaTextFieldProps, 'value' | 'onChange'> {
   className?: string;
   error?: boolean;
