+++ src/components/LoadingIndicator/LoadingIndicator.scss
