+++ src/components/LoadingIndicator/LoadingIndicator.tsx
@@ -2,13 +2,12 @@ import React, { ReactNode } from 'react';
 
 import Spinner from './Spinner';
 
-import './LoadingIndicator.module.scss';
+import './LoadingIndicator.scss';
 import classNames from 'classnames';
 
 export type LoadingIndicatorProps = {
   className?: string;
   text?: string | ReactNode;
-  style?: React.CSSProperties;
   spinnerPosition?: 'bottom' | 'left' | 'right';
 };
 
@@ -18,10 +17,9 @@ export const TEST_ID = {
   TEXT: 'loading-indicator-text'
 };
 
-const LoadingIndicator = ({ className, style, text, spinnerPosition = 'bottom' }: LoadingIndicatorProps) => {
+const LoadingIndicator = ({ className, text, spinnerPosition = 'bottom' }: LoadingIndicatorProps) => {
   return (
     <div
-      style={style}
       className={classNames('zui-loading-indicator', `zui-loading-indicator-${spinnerPosition}`, className)}
       data-testid={TEST_ID.CONTAINER}
     >
