+++ src/components/Skeleton/override.scss
@@ -9,6 +9,4 @@
 .react-loading-skeleton {
   --base-color: #{$purple-darkest};
   --highlight-color: #{$purple-lighter};
-  //--base-color: red;
-  //--highlight-color: yellow;
 }
