+++ src/components/Wizard/Presets/Confirmation.scss
@@ -6,6 +6,7 @@
 
   > p {
     text-align: center;
+    margin: 0;
 
     & p:last-of-type {
       padding-bottom: 0;
