+++ src/components/Wizard/index.ts
@@ -1,0 +1,7 @@
+import Wizard, { WizardProps, HeaderProps } from './Wizard';
+
+export type { ButtonsProps } from './Presets/Buttons';
+export type { ConfirmationProps } from './Presets/Confirmation';
+export type { LoadingProps } from './Presets/Loading';
+export type { WizardProps, HeaderProps };
+export default Wizard;
